- Renaming `animation` dsl to `transitions` - Lot more of documentation in the `kompass-android` module
