+++ android-example--fragment/src/main/java/io/sellmair/kompass/android/example/Application.kt
@@ -19,7 +19,7 @@ class Application : Application() {
         Log.d("Example", "Application.onCreate")
 
         Dependencies.router = FragmentRouter {
-            animation {
+            transitions {
                 register(LoginToLoginProcessingTransition())
                 register(ContactListToChatTransition())
                 register(LoginProcessingToLoginFailedTransition())
