+++ kompass-android/src/main/java/io/sellmair/kompass/android/ParcelableRoutingStack.kt
@@ -4,5 +4,6 @@ import android.os.Parcelable
 import io.sellmair.kompass.core.Route
 import io.sellmair.kompass.core.RoutingStack
 
+
 interface ParcelableRoutingStack<T : Route> : RoutingStack<T>, Parcelable
