--- kompass-android/src/main/java/io/sellmair/kompass/android/Parcelizer.kt
@@ -1,8 +1,0 @@
-package io.sellmair.kompass.android
-
-import android.os.Parcelable
-
-interface Parcelizer<T> {
-    fun toParcelable(value: T): Parcelable
-    fun fromParcelable(parcelable: Parcelable): T
-}
