+++ kompass-android/src/main/java/io/sellmair/kompass/android/fragment/ParcelableFragmentRouteStorage.kt
@@ -22,8 +22,7 @@ class ParcelableFragmentRouteStorage<T>(
 
     override fun get(fragment: Fragment): T {
         return getOrNull(fragment) ?: throw MissingRouteException(
-            "" +
-                    "Expected route with key $bundleKey"
+            "Expected route with key $bundleKey"
         )
     }
